refactor: combine filter/ into service/ folder
